Fix redirect_uri
